+++ platform/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.2.0] platform
 ### Releases
 - [PyPI v0.2.0] https://pypi.org/project/codat-platform/0.2.0 - platform
+
+## 2023-09-18 21:45:30
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.87.0 (2.116.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.1.0] platform
+### Releases
+- [PyPI v1.1.0] https://pypi.org/project/codat-platform/1.1.0 - platform